refactor: split bff mapping modules
